Add concept of remembering ware storage locations to the project model.
Increasingly, we're finding that project pipelines could really do with streamlining the amount of redundancy in warehouse locations.  Even more importantly, though, updating hashes for *external* upstreams that use legacy, non-content-addressable storage formats tends to be fiercely irritating, requiring both updating the hash in reppl and updating any of the formula templates that contain the URL, which is very distracting and error prone -- we need to do something to fix that.

A third call for more coherent, invisible handling of ware storage location information is the as-yet-unmade 'unpack' command.  Some reppl pipelines are actually pure: their final job is to emit some info about the product, maybe a red/green exit code, and stop.  But many many more scripts want to *do* something with that final result: unpacking it somewhere for use on the host system, perhaps.  When we do that, it should not require reiterating the storage location in the script driving reppl; we just had it in the formula, use that info!

This isn't connected to anything yet, but will be soon.

There will be tricky issues in the future with relative paths used for warehouses.  We'll cross that bridge when we come to it.  In all likelihood, those are going to be simply impossible to support well, and we'll consider changes upstream in repeatr as well to find a good solution.

Signed-off-by: Eric Myhre <hash@exultant.us>