Demo!  Of a build!  That's reproducible!  And specifies none of its warehouses explicitly; and also we can easily unpack its result afterward!  And all this works!
You can toggle which version of 'go' you're using by flipping which upstream you `reppl put` at the beginning of the demo script, and both should work correctly, without touching any of the formulas.  Both will also give you reproducible builds -- though of course different ones.  Toggling the version of go in the reppl script will bust the cache and cause a rebuild for the formula that uses it -- but not the other two, which don't depend on it, and thus (wisely) know they have no reason to be re-run!

Signed-off-by: Eric Myhre <hash@exultant.us>