Teach the 'reppl put' command to accept a kind flag.
This makes it possible to use it correctly for e.g. plain-file transport, or git transport... even for unpack.  (You could already specify the type of transport in a formula, and reppl wouldn't override it, which already made it possible to work with git etc in eval; this is also less clumsy and odd looking than that, of course, and has the database in the reppl state file less wrong.)

Signed-off-by: Eric Myhre <hash@exultant.us>